✨ feat: Implement Table of Contents component and replace SimpleTOC in BlogPostDetailUI
